fix: silence clang-tidy warnings and fix undefined behavior
make check now passes with zero warnings/errors (was 816 warnings, 14 errors):
- pass C_CONFIG to clang-tidy, run check only on .c sources, add --quiet
- include <arch/smp.h> in klogo.c (missing CPU_MAX_COUNT fallback)
- mark naked context_switch params unused via attribute instead of
  non-asm (void) statements clang rejects

Fix real UB and logic bugs:
- virtgpu execbuffer: uninitialized bos[] read when one malloc fails
- simplefs lookup: null deref after node reassigned to search result
- netlink: signed 1-bit bitfields truncate 1 to -1 (now unsigned)
- hid_parser: negative shift when report_size == 0
- extfs dir: unvalidated block_size divisor (div-by-zero), leak in
  extent_collect error paths, misplaced NOLINT lines
- tpm: uint32_t aliasing over uint8_t response buffer
- heap/socket/hda: calloc(0)/malloc(0) paths
- drm/fbcon: null derefs on driver/dirty_last_col
- make 15 memcmp/strncmp result comparisons explicit

Tune .clang-tidy for kernel-idiomatic patterns (address casts, narrowing,
void* conversions, bitmask enums, struct padding) and fix ~250 mechanical
warnings: (void) snprintf casts, switch defaults, branch-clone merges,
assignment-in-if, macro parentheses, dead stores, implicit widening

+++ drivers/ata/ahci.c
@@ -425,7 +425,7 @@ void init_ahci(void)
 
         ahci_port_state_t *port = &ahci_ports[ahci_port_count];
         memset(port, 0, sizeof(*port));
-        port->port_mmio = hba_mmio + 0x100 + i * 0x80;
+        port->port_mmio = hba_mmio + 0x100 + (size_t)i * 0x80;
         port->port_no   = (uint8_t)i;
 
         /* PI describes implemented controller ports, not attached devices.
